chore(main): Reorganise the file and directory and create a new directory
