+++ fullCourse/docs/blog11.txt
