+++ fullCourse/docs/blog2.txt
