+++ fullCourse/docs/blog3.txt
