+++ fullCourse/js/people.js
