+++ fullCourse/public/js/delete.js
