+++ fullCourse/views/details.ejs
