+++ fullCourse/views/index.ejs
