+++ fullCourse/views/partials/footer.ejs
