+++ fullCourse/views/partials/nav.ejs
